+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - TBD
+## [0.6.0] - 2023-04-23
 
 ### Added
 
@@ -146,7 +146,8 @@ Release is primarily to publish to Read the Docs as a version.
 
 - Fixed issue with item-search validation relying on collections behavior
 
-[unreleased]: https://github.com/stac-utils/stac-api-validator/compare/v0.5.0...main
+[unreleased]: https://github.com/stac-utils/stac-api-validator/compare/v0.6.0...main
+[0.6.0]: https://github.com/stac-utils/stac-api-validator/tree/v0.6.0
 [0.5.0]: https://github.com/stac-utils/stac-api-validator/tree/v0.5.0
 [0.4.3]: https://github.com/stac-utils/stac-api-validator/tree/v0.4.3
 [0.4.2]: https://github.com/stac-utils/stac-api-validator/tree/v0.4.2